requirements.txt
djangorestframework==3.12.4
kubernetes==19.15.0
psycopg2==2.9.1